requirements.txt
pandas
numpy
pymongo
seaborn
matplotlib
scikit-learn
mlflow
flask
dvc
docker
prefect
fastapi
pyyaml
python-dotenv
nltk
scipy
dill
xgboost
